Revise enabling script.

+++ tshintai_simulinkcustomtab/enable_tshintai_CustomTab.m
@@ -1,9 +1,22 @@
 %% カスタムタブを有効化
-movefile('resources', 'resources_');
-slCreateToolstripComponent("tshintai_CustomTab");
+slReloadToolstripConfig;
+CT_obj = slLoadedToolstripComponents;
+
+tab_is_not_loaded = true;
+for i = 1:numel(CT_obj)
+    if strcmp(CT_obj(i).name, 'tshintai_CustomTab')
+        tab_is_not_loaded = false;
+        break;
+    end
+end
 
 %%
-rmdir('resources', 's');
-movefile('resources_', 'resources');
-%%
-slReloadToolstripConfig;
+if tab_is_not_loaded
+    movefile('resources', 'resources_');
+    slCreateToolstripComponent("tshintai_CustomTab");
+
+    rmdir('resources', 's');
+    movefile('resources_', 'resources');
+
+    slReloadToolstripConfig;
+end
